requirements.txt
numpy==1.16.2
pygame==1.9.6
PyGLM
Pillow==5.4.1
pyrr==0.10.3
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5